Refactor transaction assembly by removing the old TransactionAssemblyStrategy and introducing new strategies for different transfer types, enhancing modularity and clarity in transaction processing.
