+++ src/main/java/info/mackiewicz/bankapp/presentation/dashboard/controller/TransferController.java
@@ -12,9 +12,9 @@ import info.mackiewicz.bankapp.account.service.AccountService;
 import info.mackiewicz.bankapp.presentation.dashboard.dto.InternalTransferRequest;
 import info.mackiewicz.bankapp.presentation.dashboard.dto.OwnTransferRequest;
 import info.mackiewicz.bankapp.presentation.dashboard.dto.TransferRequest;
-import info.mackiewicz.bankapp.transaction.controller.TransactionAssembler;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
 import info.mackiewicz.bankapp.transaction.service.TransactionService;
+import info.mackiewicz.bankapp.transaction.service.assembler.TransactionAssembler;
 import info.mackiewicz.bankapp.user.model.User;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
