+++ src/main/java/info/mackiewicz/bankapp/transaction/controller/TransactionController.java
@@ -3,7 +3,6 @@ package info.mackiewicz.bankapp.transaction.controller;
 import info.mackiewicz.bankapp.account.model.Account;
 import info.mackiewicz.bankapp.account.service.AccountService;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
-import info.mackiewicz.bankapp.transaction.model.builder.TransactionBuilder;
 import info.mackiewicz.bankapp.transaction.model.dto.CreateTransactionRequest;
 import info.mackiewicz.bankapp.transaction.service.TransactionService;
 import lombok.RequiredArgsConstructor;
@@ -19,7 +18,6 @@ import java.util.List;
 public class TransactionController {
 
     private final TransactionService transactionService;
-    private final TransactionBuilder transactionBuilder;
     private final AccountService accountService;
 
     @PostMapping
@@ -29,13 +27,13 @@ public class TransactionController {
         Account destinationAccount = request.getDestinationAccountId() != null ? 
             accountService.getAccountById(request.getDestinationAccountId()) : null;
 
-        Transaction transaction = transactionBuilder
-                .withSourceAccount(sourceAccount)
-                .withDestinationAccount(destinationAccount)
-                .withAmount(request.getAmount())
-                .withType(request.getType())
-                .withTransactionTitle(request.getTitle())
-                .build();
+        Transaction transaction = Transaction.buildTransfer()
+            .from(sourceAccount)
+            .to(destinationAccount)
+            .withAmount(request.getAmount())
+            .withTitle(request.getTitle())
+            .withTransactionType(request.getType())
+            .build();
 
         Transaction savedTransaction = transactionService.createTransaction(transaction);
         return ResponseEntity.status(HttpStatus.CREATED).body(savedTransaction);
